# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Reimbursement Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register new company and admin user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Company and admin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithCompany"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /approval-rules:
    get:
      operationId: getApprovalRules
      tags: [approval-rules]
      summary: Get approval rules for company
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Approval rules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApprovalRuleWithSteps"
    post:
      operationId: createApprovalRule
      tags: [approval-rules]
      summary: Create approval rule (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApprovalRuleRequest"
      responses:
        "201":
          description: Rule created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRuleWithSteps"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /approval-rules/{id}:
    put:
      operationId: updateApprovalRule
      tags: [approval-rules]
      summary: Update approval rule (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApprovalRuleRequest"
      responses:
        "200":
          description: Rule updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalRuleWithSteps"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteApprovalRule
      tags: [approval-rules]
      summary: Delete approval rule (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Rule deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /currencies/rates/{base}:
    get:
      operationId: getExchangeRates
      tags: [currencies]
      summary: Get exchange rates for a base currency
      parameters:
        - name: base
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Exchange rates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExchangeRates"
  /currencies/convert:
    post:
      operationId: convertCurrency
      tags: [currencies]
      summary: Convert amount between currencies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConvertCurrencyRequest"
      responses:
        "200":
          description: Converted amount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConvertCurrencyResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SignupRequest:
      type: object
      properties:
        companyName:
          type: string
        country:
          type: string
        currency:
          type: string
        adminName:
          type: string
        adminEmail:
          type: string
        adminPassword:
          type: string
      required:
        - companyName
        - country
        - currency
        - adminName
        - adminEmail
        - adminPassword
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserWithCompany"
      required:
        - token
        - user
    Company:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
        defaultCurrency:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - country
        - defaultCurrency
        - createdAt
...
    UserWithCompany:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, manager, employee]
        managerId:
          type: ["integer", "null"]
        companyId:
          type: integer
        company:
          $ref: "#/components/schemas/Company"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - companyId
        - company
        - createdAt
...
    ApprovalStep:
      type: object
      properties:
        id:
          type: integer
        ruleId:
          type: integer
        approverRole:
          type: string
          enum: [manager, admin, employee]
        approverId:
          type: ["integer", "null"]
        stepOrder:
          type: integer
      required:
        - id
        - ruleId
        - approverRole
        - stepOrder
    ApprovalRuleWithSteps:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        ruleType:
          type: string
          enum: [percentage, specific, hybrid]
        threshold:
          type: ["number", "null"]
        specificApproverId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        steps:
          type: array
          items:
            $ref: "#/components/schemas/ApprovalStep"
      required:
        - id
        - companyId
        - ruleType
        - isActive
        - createdAt
        - steps
    CreateApprovalRuleRequest:
      type: object
      properties:
        ruleType:
          type: string
          enum: [percentage, specific, hybrid]
        threshold:
          type: ["number", "null"]
        specificApproverId:
          type: ["integer", "null"]
        isActive:
          type: boolean
        steps:
          type: array
          items:
            $ref: "#/components/schemas/CreateApprovalStepRequest"
      required:
        - ruleType
        - isActive
        - steps
    CreateApprovalStepRequest:
      type: object
      properties:
        approverRole:
          type: string
          enum: [manager, admin, employee]
        approverId:
          type: ["integer", "null"]
        stepOrder:
          type: integer
      required:
        - approverRole
        - stepOrder
...
    ExchangeRates:
      type: object
      properties:
        base:
          type: string
        rates:
          type: object
          additionalProperties:
            type: number
        updatedAt:
          type: string
          format: date-time
      required:
        - base
        - rates
        - updatedAt
    ConvertCurrencyRequest:
      type: object
      properties:
        amount:
          type: number
        fromCurrency:
          type: string
        toCurrency:
          type: string
      required:
        - amount
        - fromCurrency
        - toCurrency
    ConvertCurrencyResponse:
      type: object
      properties:
        amount:
          type: number
        fromCurrency:
          type: string
        toCurrency:
          type: string
        convertedAmount:
          type: number
        rate:
          type: number
      required:
        - amount
        - fromCurrency
        - toCurrency
        - convertedAmount
        - rate
...