# lib/api-zod/src/generated/types/convertCurrencyResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reimbursement Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface ConvertCurrencyResponse {
  amount: number;
  fromCurrency: string;
  toCurrency: string;
  convertedAmount: number;
  rate: number;
}